+++ assets/stylesheets/main.scss
@@ -63,7 +63,7 @@
 @import '../pages/contributions-landing/contributionsLanding';
 @import '../pages/subscriptions-landing/subscriptionsLanding';
 @import '../pages/regular-contributions/regularContributions';
-@import '../components/contributionsThankYou/contributionsThankYouPage';
+@import '../containerisableComponents/contributionsThankYou/contributionsThankYouPage';
 @import '../pages/oneoff-contributions/oneoffContributions';
 @import '../pages/regular-contributions-existing/regularContributionsExisting';
 @import '../pages/paypal-error/payPalError';
